requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# CLIP model
git+https://github.com/openai/CLIP.git

# Additional utilities
tqdm>=4.62.0
opencv-python>=4.5.0
albumentations>=1.3.0
grad-cam>=1.4.0

# Development and reproducibility
pytest>=6.2.0
black>=21.0.0
flake8>=4.0.0
tensorboard>=2.8.0

# Optional: For advanced visualization
plotly>=5.0.0
ipywidgets>=7.6.0